fix discovered artifact panic
When known discoveries was displayed on a window of type NHW_MENU,
header lines sent to it via menu_add_heading() disappeared into the
bit bucket.  Switching back to putstr() made them show up.  But I also
changed the type to NHW_TEXT.  A menu_add_heading() in artifact.c was
overlooked, and after the window type change that started triggering
a panic when '\' or '`a' was used while any artifacts were discovered.

Not sure how other interfaces were dealing with this.