unfix unnaming types of objects
The prior commit makes the code clearer so I'm not reverting it,
but the old code was not accessing freed memory so I am removing the
fixes entry.  The static analyzer's complaint is bogus.  Freeing the
memory that held a user-assigned type name did not affect whether
that pointer could be tested for being Null.  Its stale value wasn't
being dereferenced.