more PR #1193 - doors
The old default allocation of space for 120 doors on a level seems
excessive.  Reduce that to 20.  Code for expanding the allocation
when needed is already in place.

Pull request from #1193 from mkuoppal was superseded by commits
210d5d5c63564008a5e3b581515cbbd3a99fe6ce and
aeecf5f35c57bbbc4eb9c966d1fa038b60590f45 so mark it closed.

Closes #1193