Forcefighting webs (when adjacent, not trapped)
ACURR(A_STR) can yield a value as high as 125.  Switch to acurrstr()-2
for a value of 1..19.

P_SKILL(skilltype) shouldn't be used unconditionally.  It yields a
different value for restricted than for unskilled and those two skill
levels only differ for training, not effectiveness.

Allow weaponless forcefight againsta webs.  The chance for success may
need to be tweaked.