loadstone bit
If your inventory is full and you aren't already carrying a loadstone,
you can pick one up into the overflow slot.  But if you are already
carrying one and the one you're trying to pick up won't merge with it
(only criterium that matters would be BUC state, I think), you can't
pick it up and get a message saying so.  If loadstone isn't known
yet, the message always referred to it as "gray stone" rather than
"stone called <whatever-you-called-it>".