fix "It stole <item>" if hero teleports to safety
Having an item that allowed being in/on water/lava be stolen could
result in the hero teleporting.  If the thief wasn't visible from
hero's new location, the message given was "It stole <whatever>."

Save the thief's name/description before removing the item and
potentially dropping hero into trouble and triggering teleport, then
use that in the eventual message.