fix github issue #1204 - reflected poison gas
Reported by elunna:  a poison gas breath attack that was reflected
by the target still left that target enveloped in a poison gas cloud.

This makes the gas trail not extend to the target if the attack hits
and is reflected.  But if the attack misses then the cloud does reach
the target, which seems weird to me.  However, being in the cloud is
a separate event that isn't deterred by reflection.

Closes #1204