theft against punished hero with empty inventory
If a punished hero's inventory is empty, have a nymph usually take
off the attached chain and indirectly fix punishment.  If hero is
trapped by being chained to a buried iron ball, sometimes take off
the implicit chain and indirectly release hero from the trap.

Monkeys won't do this and nymphs will only do so if there is no
inventory (aside from gold and/or embedded dragon scales which they
aren't allowed to target).

It's a lot simpler than I was expecting when I wrote the TODO about
it recently.