more nhmd4.c formatting
Insert a space between cast operator and target.

For 'x | y' expression spanning multiple lines, split with bitwise OR
operator at the beginning of second line instead at the end of first.

I changed a macro that was expanding to 'expr_1; expr_2' into
'(expr_1, expr_2)'.  Both expressions modify arg 'a' which would be
a no-no within a single expression, but the comma operator provides a
sequence point that guarantees that the first expression will be fully
evaluated with side-effects--assignment to 'a'--completed before the
second one starts, making both assignments be well defined.