fix #4131 - medusa-4 level layout
The medusa-3 and medusa-4 levels each have four possible locations
for Medusa.  medusa-3 forced the downstairs to have the same location
but medusa-4 was choosing Medusa's location independently from the
stair location.  So it was likly that the hero could go down to the
next level without having to deal with her.