Sunsword vs <u.ux,u.uy>
Using #invoke on wielded or carried Sunsword and picking direction
'>' or '<' lights the hero's spot.  But setting levl[u.ux][u.uy].lit
was too simplistic.  Lighting on the Rogue Level operates on full
rooms when done inside a room and doesn't to anything with done in a
corridor.

litroom() gave inappropriate messages in a couple of special cases.